// apps/api/src/modules/booking/entities/booking.entity.ts
export enum BookingType {
  DAILY = 'DAILY',
  HOURLY = 'HOURLY',
}

export enum BookingSource {
  DIRECT = 'DIRECT',
  BLUESTAY = 'BLUESTAY',
  WALK_IN = 'WALK_IN',
}

export enum BookingStatus {
  PENDING = 'PENDING',
  CONFIRMED = 'CONFIRMED',
  CHECKED_IN = 'CHECKED_IN',
  CHECKED_OUT = 'CHECKED_OUT',
  CANCELLED = 'CANCELLED',
  NO_SHOW = 'NO_SHOW',
}

export enum PaymentStatus {
  PENDING = 'PENDING',
  PAID = 'PAID',
  PARTIALLY_REFUNDED = 'PARTIALLY_REFUNDED',
  REFUNDED = 'REFUNDED',
  FAILED = 'FAILED',
}